% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RrH.r
\name{RrH.mad}
\alias{RrH.mad}
\title{Robust estimator for a generalized ratio model
with Huber's weight function and MAD scal
by iteratively re-weighted least squares (IRLS) algorithm for M-estimation}
\usage{
RrH.mad(x1, y1, g1 = 0.5, c1 = 2.88, rp.max = 100, cg.rt = 0.01)
}
\arguments{
\item{x1}{single explanatory variable}

\item{y1}{objective variable}

\item{g1}{power (default: g1=0.5(conventional ratio model))}

\item{c1}{tuning parameter usually from 1.44 to 2.88 (equivalent to those for AAD scale)}

\item{rp.max}{maximum number of iteration}

\item{cg.rt}{convergence condition to stop iteration (default: cg1=0.001)}
}
\value{
a list with the following elements
par   robustly estimated ratio of y1 to x1
res   homoscedastic quasi-residuals
wt    robust weights
rp    total number of iteration
s1    changes of the scale (AAD or MAD)
efg   error flag. 1: acalculia (all weights become zero)  0: successful termination
}
\description{
Robust estimator for a generalized ratio model
with Huber's weight function and MAD scal
by iteratively re-weighted least squares (IRLS) algorithm for M-estimation
}
