% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RrH.r
\name{RrHa.aad}
\alias{RrHa.aad}
\title{RrHa.aad: Robust estimator for a generalized ratio model (gamma=1)
with Huber's weight function and AAD scal
by iteratively re-weighted least squares (IRLS) algorithm for M-estimation}
\usage{
RrHa.aad(x1, y1, c1 = 2.3, rp.max = 100, cg.rt = 0.01)
}
\arguments{
\item{x1}{single explanatory variable}

\item{y1}{objective variable}

\item{c1}{tuning parameter from 1.15 to 2.30 for the scale parameter of AAD(average absolute deviation)
and from 1.44 to 2.88 for the scale parameter of MAD(median absolute deviation)}

\item{rp.max}{maximum number of iteration}

\item{cg.rt}{convergence condition to stop iteration (default: cg1=0.001)}
}
\value{
a list with the following elements
par   robustly estimated ratio of y1 to x1
res   homoscedastic quasi-residuals
wt    robust weights
rp    total number of iteration
s1    changes of the scale (AAD or MAD)
efg   error flag. 1: acalculia (all weights become zero)  0: successful termination
}
\description{
RrHa.aad: Robust estimator for a generalized ratio model (gamma=1)
with Huber's weight function and AAD scal
by iteratively re-weighted least squares (IRLS) algorithm for M-estimation
}
