% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RrT.r
\name{RrTb.mad}
\alias{RrTb.mad}
\title{RrTb.mad: Robust estimator for conventional ratio model (gamma=1/2)
with Tukey biweight function and MAD scale
by iteratively re-weighted least squares (IRLS) algorithm for M-estimation}
\usage{
RrTb.mad(x1, y1, c1 = 10.03, rp.max = 100, cg.rt = 0.01)
}
\arguments{
\item{x1}{single explanatory variable}

\item{y1}{objective variable}

\item{c1}{tuning parameter from 4 to 8 for the scale parameter of AAD(average absolute deviation)
and from 5.01 to 10.03 for the scale parameter of MAD(median absolute deviation)}

\item{rp.max}{maximum number of iteration}

\item{cg.rt}{convergence condition to stop iteration (default: cg1=0.001)}
}
\value{
a list with the following elements
par   robustly estimated ratio of y1 to x1
res   homoscedastic quasi-residuals
wt    robust weights
rp    total number of iteration
s1    changes of the scale (AAD or MAD)
efg   error flag. 1: acalculia (all weights become zero)  0: successful termination
}
\description{
RrTb.mad: Robust estimator for conventional ratio model (gamma=1/2)
with Tukey biweight function and MAD scale
by iteratively re-weighted least squares (IRLS) algorithm for M-estimation
}
