% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tirls.r
\name{Tirls.aad}
\alias{Tirls.aad}
\title{Tirls.aad: Robust estimator for the linear regression model
with Tukey's biweight function and AAD scal
by iteratively re-weighted least squares (IRLS) algorithm}
\usage{
Tirls.aad(x1, y1, rt = rep(1, length(y1)), c1 = 8, rp.max = 150, cg.rt = 0.01)
}
\arguments{
\item{x1}{explanatory variable(s)}

\item{y1}{objective variable}

\item{rt}{sample weights}

\item{c1}{tuning parameter from 4 to 8 for the scale parameter of AAD(average absolute deviation)}

\item{rp.max}{maximum number of iteration}

\item{cg.rt}{convergence condition to stop iteration (default: cg1=0.001)}
}
\value{
a list with the following elements
\describe{
\item{\code{TK}}{results of robust regression}
\item{\code{wt}}{robust weights}
\item{\code{rp}}{total number of iteration}
\item{\code{s1}}{changes of the scale (AAD or MAD)}
}
}
\description{
Tirls.aad: Robust estimator for the linear regression model
with Tukey's biweight function and AAD scal
by iteratively re-weighted least squares (IRLS) algorithm
}
