% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tirls.r
\name{Tirls.mad}
\alias{Tirls.mad}
\title{Tirls.mad: Robust estimator for the linear regression model
with Tukey's biweight function and MAD scale
by iteratively re-weighted least squares (IRLS) algorithm}
\usage{
Tirls.mad(
  x1,
  y1,
  rt = rep(1, length(y1)),
  c1 = 10.03,
  rp.max = 150,
  cg.rt = 0.01
)
}
\arguments{
\item{x1}{explanatory variable(s)}

\item{y1}{objective variable}

\item{rt}{sample weights}

\item{c1}{tuning parameter from 5.01 to 10.03 for the scale parameter of MAD(median absolute deviation)}

\item{rp.max}{maximum number of iteration}

\item{cg.rt}{convergence condition to stop iteration (default: cg1=0.001)}
}
\value{
a list with the following elements
\describe{
\item{\code{TK}}{results of robust regression}
\item{\code{wt}}{robust weights}
\item{\code{rp}}{total number of iteration}
\item{\code{s1}}{changes of the scale (AAD or MAD)}
}
}
\description{
Tirls.mad: Robust estimator for the linear regression model
with Tukey's biweight function and MAD scale
by iteratively re-weighted least squares (IRLS) algorithm
}
