% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robReg.r
\name{robReg}
\alias{robReg}
\title{\code{robReg} is for Robust regression by the IRLS algorithm.
It integrates Tirls.r and Hirls.r.}
\usage{
robReg(
  x1,
  y1,
  wf = "T",
  scale = "AAD",
  rt = rep(1, length(y1)),
  tp = 8,
  rp.max = 150,
  cg.rt = 0.01
)
}
\arguments{
\item{x1}{explanatory variable in regression (a vector or a matrix)}

\item{y1}{objective variable in regression (a vector)}

\item{wf}{weight function ("T" for Tukey's biweight, and "H" for Huber weight)}

\item{scale}{scale for residuals. "AAD"(default) or "MAD".}

\item{rt}{sample weight (default 1)}

\item{tp}{tuning parameter (tp=4, 6 or 8) for weight function.
Smaller figure is more robust.}

\item{rp.max}{The maximum number of iteration (default 150)}

\item{cg.rt}{convergence condition to stop iteration (default: cg.rt=0.001)}
}
\value{
a list with the following elements
\describe{
\item{\code{cond}}{Weight function, scale, and other arguments choosed}
\item{\code{TK}}{robustly estimated regression coefficients using Tukey's biweight}
\item{\code{HB}}{robustly estimated regression coefficients using Huber weight}
\item{\code{wt}}{final robust weights}
\item{\code{rp}}{total number of iteration}
\item{\code{s1}}{iterative changes in the sclae of residuals (AAD or MAD)}
}
}
\description{
\code{robReg} is for Robust regression by the IRLS algorithm.
It integrates Tirls.r and Hirls.r.
}
